% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boarding_school_flu}
\alias{boarding_school_flu}
\title{Boarding-school influenza, England, 1978}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{time}{Day of the outbreak, 0 = the index case.}
  \item{sInfected}{Boys confined to bed.}
}
}
\source{
Anonymous (1978) "Influenza in a boarding school",
  *British Medical Journal* 1:587.
}
\usage{
boarding_school_flu
}
\description{
Daily count of boys confined to bed during the influenza outbreak at an
English boarding school, reported in the *British Medical Journal*,
4 March 1978. 763 boys, one index case, about two weeks of counts. This is
the series fitted in `sd_example("boarding_school_flu")`.
}
\keyword{datasets}
