% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{meta}
\alias{meta}
\title{Model metadata}
\usage{
meta(name = NULL, ...)
}
\arguments{
\item{name}{Model name, used in printing and as the default plot title.}

\item{...}{Further free-form metadata (author, notes, source, ...).}
}
\value{
An `sd_element` to be passed to [sd_structure()].
}
\description{
Model metadata
}
