% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{sd_example_run}
\alias{sd_example_run}
\title{Run a catalogue model}
\usage{
sd_example_run(id, ...)
}
\arguments{
\item{id}{A number 1-22 or the model's name.}

\item{...}{Passed to [simulate()], overriding the example's own arguments.}
}
\value{
An `sd_result` tibble.
}
\description{
Run a catalogue model
}
\examples{
out <- sd_example_run("teacup")
}
