% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sd_parameters}
\alias{sd_parameters}
\title{The parameter layer: the numbers and the data}
\usage{
sd_parameters(...)
}
\arguments{
\item{...}{Parameter elements.}
}
\value{
An object of class `sd_parameters`.
}
\description{
`sd_parameters()` collects [constant()], [reference()], [initial()],
[lookup_data()] and [input_series()] declarations.
}
\examples{
sd_parameters(
  constant(contact_rate = 6, infectivity = 0.2),
  initial(S = 999, I = 1, R = 0)
)
}
