% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysd-package.R
\docType{package}
\name{tidysd-package}
\alias{tidysd}
\alias{tidysd-package}
\title{tidysd: a tidy grammar for system dynamics}
\description{
Specify a stock-and-flow model as three independent layers and simulate it
with one verb:
}
\details{
\preformatted{
sd_structure()    what exists, how it is wired      (no math)
sd_equations()    functional forms                  (symbolic)
sd_parameters()   the numbers and data

simulate(structure, equations, parameters, spec = sim_spec())
}

There is no compile step: binding and validation happen inside
[simulate()]. The "model" is just the three layer objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rayhanalirachman.github.io/tidySD/}
  \item \url{https://github.com/rayhanalirachman/tidySD}
  \item Report bugs at \url{https://github.com/rayhanalirachman/tidySD/issues}
}

}
\author{
\strong{Maintainer}: Rayhan Ali Rachman \email{rayhanalirachman@gmail.com}

Authors:
\itemize{
  \item Rayhan Ali Rachman \email{rayhanalirachman@gmail.com}
}

}
\keyword{internal}
