% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{update.sd_equations}
\alias{update.sd_equations}
\title{Replace or add equations}
\usage{
\method{update}{sd_equations}(object, ...)
}
\arguments{
\item{object}{An `sd_equations` object.}

\item{...}{Two-sided formulas.}
}
\value{
A new `sd_equations` object.
}
\description{
Layers are immutable: `update()` returns a new `sd_equations` object with the
given formulas replacing (by left-hand side) or extending the existing ones.
}
\examples{
e <- sd_equations(lambda ~ beta * I, IR ~ S * lambda)
update(e, lambda ~ contact_rate * infectivity * I)
}
