% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{update.sd_parameters}
\alias{update.sd_parameters}
\title{Replace or add parameters}
\usage{
\method{update}{sd_parameters}(object, ...)
}
\arguments{
\item{object}{An `sd_parameters` object.}

\item{...}{Parameter elements, or named numeric values (treated as constants).}
}
\value{
A new `sd_parameters` object.
}
\description{
Replace or add parameters
}
\examples{
p <- sd_parameters(constant(a = 1), initial(S = 10))
update(p, constant(a = 2))
update(p, a = 3)
}
