% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{autoplot.sd_diagram}
\alias{autoplot.sd_diagram}
\alias{plot.sd_diagram}
\title{Plot a model diagram}
\usage{
\method{autoplot}{sd_diagram}(object, ...)

\method{plot}{sd_diagram}(x, ...)
}
\arguments{
\item{object}{An [sd_diagram()].}

\item{...}{Unused.}

\item{x}{An [sd_diagram()].}
}
\value{
A `ggplot` object.
}
\description{
A deliberately plain layered layout: stocks on the spine, everything else
arranged around them. It is meant for reading a model, not for publication.
}
